+++ highcharts_stock/options/series/data/treegraph.py
@@ -1,0 +1,1 @@
+from highcharts_core.options.series.data.treegraph import *